[MISC] move seqan3::views::minimiser_hash to seqan3/search/views/
