+++ doc/tutorial/search/views/minimiser/minimiser_solution2.cpp
@@ -1,6 +1,6 @@
 #include <seqan3/alphabet/nucleotide/dna4.hpp>
 #include <seqan3/core/debug_stream.hpp>
-#include <seqan3/range/views/minimiser_hash.hpp>
+#include <seqan3/search/views/minimiser_hash.hpp>
 
 using seqan3::operator""_dna4;
 using seqan3::operator""_shape;
