+++ test/performance/range/views/CMakeLists.txt
@@ -1,8 +1,6 @@
 seqan3_benchmark(view_all_benchmark.cpp)
 seqan3_benchmark(view_drop_benchmark.cpp)
 seqan3_benchmark(view_drop_view_take_benchmark.cpp)
-seqan3_benchmark(view_kmer_hash_benchmark.cpp)
-seqan3_benchmark(view_minimiser_hash_benchmark.cpp)
 seqan3_benchmark(view_take_benchmark.cpp)
 seqan3_benchmark(view_take_until_benchmark.cpp)
 seqan3_benchmark(view_translate_1D_benchmark.cpp)
