+++ test/performance/search/views/view_minimiser_hash_benchmark.cpp
@@ -8,7 +8,7 @@
 #include <benchmark/benchmark.h>
 
 #include <seqan3/alphabet/nucleotide/dna4.hpp>
-#include <seqan3/range/views/minimiser_hash.hpp>
+#include <seqan3/search/views/minimiser_hash.hpp>
 #include <seqan3/test/performance/naive_minimiser_hash.hpp>
 #include <seqan3/test/performance/sequence_generator.hpp>
 #include <seqan3/test/performance/units.hpp>
