+++ test/unit/search/views/CMakeLists.txt
@@ -1,2 +1,3 @@
 seqan3_test (kmer_hash_test.cpp)
+seqan3_test (minimiser_hash_test.cpp)
 seqan3_test (minimiser_test.cpp)
